move generator loading and serialization into transform module
