+++ libs/datamodel/core/src/transform/ast_to_dml/mod.rs
@@ -1,4 +1,5 @@
 mod common;
+mod generator_loader;
 mod invalid_model_names;
 mod lift;
 mod precheck;
@@ -6,6 +7,7 @@ mod standardise;
 mod validate;
 mod validation_pipeline;
 
+pub use generator_loader::GeneratorLoader;
 use lift::*;
 use standardise::*;
 use validate::*;
