+++ libs/datamodel/core/src/transform/dml_to_ast/mod.rs
@@ -1,3 +1,5 @@
+mod generator_serializer;
 mod lower;
 
+pub use generator_serializer::GeneratorSerializer;
 pub use lower::LowerDmlToAst;
